requirements.txt
flask==2.0.1
flask-cors==3.0.10
pyotp==2.6.0
python-dotenv==0.19.0